fix: move getUserInfo to Protected route HOC
